requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torch-fidelity>=0.3.0
torchmetrics>=1.8.2
tqdm>=4.67.1
numpy>=1.24.0
scipy>=1.10.0
nibabel>=5.0.0
PyYAML>=6.0
tqdm>=4.65.0
scikit-learn>=1.3.0
einops>=0.8.1
fidelity>=0.1
ImageIO>=2.37.2
opencv-python>=4.12.0.88
scikit-image>=0.25.2
scikit-learn>=1.7.2
scipy>=1.16.2
matplotlib>=3.10.7
matplotlib-inline>=0.1.7
